+++ db/migrate/20260804000001_add_judge_temperature_ignored_to_runs.rb
@@ -1,0 +1,5 @@
+class AddJudgeTemperatureIgnoredToRuns < ActiveRecord::Migration[8.0]
+  def change
+    add_column :completion_kit_runs, :judge_temperature_ignored, :boolean, default: false, null: false
+  end
+end